--- a/barretenberg/cpp/src/barretenberg/honk/execution_trace/execution_trace_usage_tracker.test.cpp
+++ b/barretenberg/cpp/src/barretenberg/honk/execution_trace/execution_trace_usage_tracker.test.cpp
@@ -30,28 +30,12 @@ TEST_F(ExecutionTraceUsageTrackerTest, ConstructThreadRanges)
 {
     using Range = ExecutionTraceUsageTracker::Range;
 
-    std::vector<Range> union_ranges = { { 2, 8 }, { 13, 34 }, { 36, 42 }, { 50, 57 } };
+    std::vector<Range> union_ranges = { { 2, 8 }, { 13, 34 }, { 36, 45 }, { 50, 60 } };
 
-    std::vector<std::vector<Range>> expected_thread_ranges = {
-        { { 2, 8 }, { 13, 17 } }, { { 17, 27 } }, { { 27, 34 }, { 36, 39 } }, { { 39, 42 }, { 50, 57 } }
-    };
-
-    const size_t num_threads = 4;
-    std::vector<std::vector<Range>> thread_ranges =
-        ExecutionTraceUsageTracker::construct_ranges_for_equal_content_distribution(union_ranges, num_threads);
-
-    EXPECT_EQ(thread_ranges, expected_thread_ranges);
-}
-
-TEST_F(ExecutionTraceUsageTrackerTest, ConstructThreadRangesNotDivisible)
-{
-    using Range = ExecutionTraceUsageTracker::Range;
-
-    std::vector<Range> union_ranges = { { 2, 8 }, { 13, 34 }, { 36, 42 }, { 50, 60 } };
-
-    std::vector<std::vector<Range>> expected_thread_ranges = {
-        { { 2, 8 }, { 13, 17 } }, { { 17, 27 } }, { { 27, 34 }, { 36, 39 } }, { { 39, 42 }, { 50, 60 } }
-    };
+    std::vector<std::vector<Range>> expected_thread_ranges = { { { 2, 4 }, { 13, 19 }, { 36, 39 }, { 50, 53 } },
+                                                               { { 4, 6 }, { 19, 24 }, { 39, 41 }, { 53, 56 } },
+                                                               { { 6, 7 }, { 24, 29 }, { 41, 43 }, { 56, 58 } },
+                                                               { { 7, 8 }, { 29, 34 }, { 43, 45 }, { 58, 60 } } };
 
     const size_t num_threads = 4;
     std::vector<std::vector<Range>> thread_ranges =
@@ -66,7 +50,7 @@ TEST_F(ExecutionTraceUsageTrackerTest, ConstructThreadRangesMoreThreadsThanWork)
 
     std::vector<Range> union_ranges = { { 2, 3 }, { 13, 14 } };
 
-    std::vector<std::vector<Range>> expected_thread_ranges = { {}, {}, {}, { { 2, 3 }, { 13, 14 } } };
+    std::vector<std::vector<Range>> expected_thread_ranges = { { { 2, 3 }, { 13, 14 } }, {}, {}, {} };
 
     const size_t num_threads = 4;
     std::vector<std::vector<Range>> thread_ranges =
@@ -83,9 +67,9 @@ TEST_F(ExecutionTraceUsageTrackerTest, ConstructThreadRangesSplitsLargeRange)
     // Single range that is too large to fit in a single thread
     std::vector<Range> union_ranges = { { 0, 1 }, { 2, 101 } };
 
-    std::vector<std::vector<Range>> expected_thread_ranges = { { { 0, 1 }, { 2, 34 } },
-                                                               { { 34, 67 } },
-                                                               { { 67, 101 } } };
+    std::vector<std::vector<Range>> expected_thread_ranges = { { { 0, 1 }, { 2, 35 } },
+                                                               { { 35, 68 } },
+                                                               { { 68, 101 } } };
 
     const size_t num_threads = 3;
     std::vector<std::vector<Range>> thread_ranges =