chrome: Distribute content equally among threads within each range

The original code gives each thread the same number of rows but the same number of rows in each range does not necessarily mean the same work. This PR makes sure that for each range, each thread gets roughly the same number of rows. The leftovers are also evenly distributed because if we give them to the last thread for each range, the total worst case the last thread can get O(num_thread * num_ranges) more rows than others.