@johnathan79717 2025-06-30
I ran the benchmark on the mainframe 3 times with this commit applied and unapplied alternatively. The results are not super consistent and I guess it may be affected by the mainframe load at the moment. Is there a more reliable way to benchmark it?

Round 1

Before:
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
ClientIVCBench/Full/6      23006 ms        19089 ms            1

After
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
ClientIVCBench/Full/6      20792 ms        16438 ms            1

Round 2

Before:
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
ClientIVCBench/Full/6      20199 ms        16235 ms            1

After
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
ClientIVCBench/Full/6      20655 ms        16549 ms            1

Round 3

Before
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
ClientIVCBench/Full/6      20677 ms        16355 ms            1

After
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
ClientIVCBench/Full/6      19577 ms        15720 ms            1