@johnathan79717 2025-06-30
Tracy with the `deploy_ecdsar1+sponsored_fpc` flow:

Before:

First `prove` round:
<img width="3472" alt="image" src="https://github.com/user-attachments/assets/e3c23e26-da82-4fe6-84dc-975bee22b34f" />

After
<img width="3472" alt="image" src="https://github.com/user-attachments/assets/f109262c-aa7a-4faa-a3ce-d12d630cd48b" />

Last `prove` round:
Before:
<img width="3472" alt="image" src="https://github.com/user-attachments/assets/67da0b44-0f93-4d41-9c08-e39fa0f902fa" />

After:
<img width="3472" alt="image" src="https://github.com/user-attachments/assets/fab14f3f-5dea-47f0-88a1-95a0ae022323" />

From the above, it does seem that the work load is more balanced.