@lucasxia01 2025-06-30
benchmark_remote.sh should be able to run your benchmark on a separate machine.