@ledwards2225 2025-07-01
Its a bit hard to tell what the actual effect is from the tracy graphs. E.g. in the second chart the combiner appears to have a more balanced thread distribution but the total computation time appears to be significantly greater. I suppose in addition to the times being distorted within a single run they probably can't be compared between runs either.

As Lucas suggested, using the script `benchmark_remote.sh` should allow you to run on designated resources for better consistency. In addition to the top level numbers it would be good to see the specific numbers for the perturbator/combiner computations. Hopefully they go down!