@johnathan79717 2025-07-01
I haven't been able to use `benchmark_remote.sh`, but I was able to get specific numbers for perturbator/combiner using 
```
CI=1 NO_WASM=1 NATIVE_PRESET=op-count-time ./barretenberg/cpp/bootstrap.sh bench_ivc deploy_ecdsar1+sponsored_fpc
```
on mainframe.

Again, I alternated between "With change" and "Without change".

### 1st run with change

function          |                        ms
-|-
ProtogalaxyProver::prove(t)      |       8520
compute_combiner(t)              |       1595 
compute_perturbator(t)           |       1101 
Total time | 22389 

### 1st run without change:

function          |                        ms
-|-
ProtogalaxyProver::prove(t)      |       8218
compute_combiner(t)              |       1713 
compute_perturbator(t)           |       1026 
Total time | 22040

### 2nd run with change

function          |                        ms
-|-
ProtogalaxyProver::prove(t)      |       7717
compute_combiner(t)              |       1314 
compute_perturbator(t)           |       995 
Total time | 21114


#### 2nd run without change

function          |                        ms
-|-
ProtogalaxyProver::prove(t)      |       7995
compute_combiner(t)              |       1661 
compute_perturbator(t)           |       1027 
Total time | 21474

### 3rd run with change

function          |                        ms
-|-
ProtogalaxyProver::prove(t)      |       7642
compute_combiner(t)              |       1291 
compute_perturbator(t)           |       1045 
Total time | 26254

### 3rd run without change

function          |                        ms
-|-
ProtogalaxyProver::prove(t)      |       8724
compute_combiner(t)              |       1735 
compute_perturbator(t)           |       1125 
Total time | 23720

Even though this is not a large sample and there's still variance with running on the mainframe. It seems like `compute_combiner` is consistently faster with this change than without.