@ledwards2225 2025-07-01
I was able to run next vs branch on the designated resources via `benchmark_client_ivc.sh` and the results indicate a negligible difference (likely within noise):

next:
```
ClientIVCBench/Full/6      26687 ms        22385 ms
ProtogalaxyProver::prove(t)            12070    48.05%
compute_combiner(t)                     3798   15.12%
compute_perturbator(t)                  2513   10.01%
```

branch:
```
ClientIVCBench/Full/6      26449 ms        22666 ms
ProtogalaxyProver::prove(t)            11900    47.77%
compute_combiner(t)                     3627   14.56%
compute_perturbator(t)                  2482    9.96%
```

Its definitely worth performing this comparison on the designated resources for some of the real transaction flows since this basic benchmark (Full6) is a bit contrived. If we don't see anything significant there, it may be time to close this chapter, at least for the time being.