@johnathan79717 2025-07-02
Similar to `benchmark_client_ivc.sh`, I wrote a [script](https://github.com/AztecProtocol/aztec-packages/pull/15469) to run example flows on the remote machine. I'll post results soon.